# doc/swagger.yaml
definitions:
...
  github_com_cyware_ssi-service_pkg_service_webhook.Noun:
    enum:
    - Credential
    - DID
    - Manifest
    - SchemaID
    - Presentation
    - Application
    - Submission
    type: string
    x-enum-varnames:
    - Credential
    - DID
    - Manifest
    - Schema
    - Presentation
    - Application
    - Submission
  github_com_cyware_ssi-service_pkg_service_webhook.Verb:
    enum:
    - BatchCreate
    - Create
    - Delete
    type: string
    x-enum-varnames:
    - BatchCreate
    - Create
    - Delete
  github_com_cyware_ssi-service_pkg_service_webhook.Webhook:
    properties:
      noun:
        $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_webhook.Noun'
      urls:
        items:
          type: string
        type: array
      verb:
        $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_webhook.Verb'
    required:
    - noun
    - urls
    - verb
    type: object
...
  pkg_server_router.CreateSchemaRequest:
    properties:
      description:
        description: Description is an optional human-readable description for a schema
        type: string
      issuer:
        description: |-
          Issuer represents the DID of the issuer for the schema if it's signed. Required if intending to sign the
          schema as a credential using JsonSchemaCredential.
        type: string
      name:
        description: Name is a human-readable name for a schema
        type: string
      schema:
        allOf:
        - $ref: '#/definitions/schema.JSONSchema'
        description: |-
          Schema represents the JSON schema for the credential schema
          If the schema has an $id field, it will be overwritten with an ID the service generates.
          The schema must be against draft 2020-12, 2019-09, or 7.
          Must include a string field `$schema` that must be one of `https://json-schema.org/draft/2020-12/schema`,
          `https://json-schema.org/draft/2019-09/schema`, or `https://json-schema.org/draft-07/schema`.
      verificationMethodId:
        description: |-
          The id of the verificationMethod (see https://www.w3.org/TR/did-core/#verification-methods) who's privateKey is
          stored in ssi-service. The verificationMethod must be part of the did document associated with `issuer`.
          The private key associated with the verificationMethod's publicKey will be used to sign the schema as a
          credential using JsonSchemaCredential.
          Required if intending to sign the schema as a credential using JsonSchemaCredential.
        example: did:key:z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3#z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3
        type: string
    required:
    - issuer
    - name
    - schema
    - verificationMethodId
    type: object
  pkg_server_router.CreateSchemaResponse:
    properties:
      credentialSchema:
        description: CredentialSchema is the JWT schema for the credential, returned
          when the type is CredentialSchema
        type: string
      id:
        description: ID is the URL of for resolution of the schema
        type: string
      schema:
        allOf:
        - $ref: '#/definitions/schema.JSONSchema'
        description: Schema is the JSON schema for the credential, returned when the
          type is JsonSchema
      type:
        allOf:
        - $ref: '#/definitions/schema.VCJSONSchemaType'
        description: Type is the type of schema such as `JsonSchema` or `JsonSchemaCredential`
    required:
    - type
    type: object
...
  pkg_server_router.CreateWebhookRequest:
    properties:
      noun:
        allOf:
        - $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_webhook.Noun'
        description: 'The noun (entity) for the new webhook.eg: Credential'
      url:
        description: The URL to post the output of this request to Noun.Verb action
          to.
        type: string
      verb:
        allOf:
        - $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_webhook.Verb'
        description: 'The verb for the new webhook.eg: Create'
    required:
    - noun
    - url
    - verb
    type: object
  pkg_server_router.CreateWebhookResponse:
    properties:
      webhook:
        $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_webhook.Webhook'
    type: object
...
  pkg_server_router.GetSchemaResponse:
    properties:
      credentialSchema:
        description: CredentialSchema is the JWT schema for the credential, returned
          when the type is CredentialSchema
        type: string
      id:
        description: ID is the URL of for resolution of the schema
        type: string
      schema:
        allOf:
        - $ref: '#/definitions/schema.JSONSchema'
        description: Schema is the JSON schema for the credential, returned when the
          type is JsonSchema
      type:
        allOf:
        - $ref: '#/definitions/schema.VCJSONSchemaType'
        description: Type is the type of schema such as `JsonSchema` or `JsonSchemaCredential`
    required:
    - type
    type: object
...
  pkg_server_router.ListSchemasResponse:
    properties:
      nextPageToken:
        description: Pagination token to retrieve the next page of results. If the
          value is "", it means no further results for the request.
        type: string
      schemas:
        description: Schemas is the list of all schemas the service holds
        items:
          $ref: '#/definitions/pkg_server_router.GetSchemaResponse'
        type: array
    type: object
...
  pkg_server_router.ListWebhookResponse:
    properties:
      webhook:
        $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_webhook.Webhook'
    type: object
  pkg_server_router.ListWebhooksResponse:
    properties:
      webhooks:
        items:
          $ref: '#/definitions/pkg_server_router.ListWebhookResponse'
        type: array
    type: object
...
  schema.JSONSchema:
    additionalProperties: {}
    type: object
  schema.VCJSONSchemaType:
    enum:
    - JsonSchemaCredential
    - JsonSchema
    type: string
    x-enum-varnames:
    - JSONSchemaCredentialType
    - JSONSchemaType
...
info:
  contact:
    email: tbd-developer@squareup.com
    name: CYWARE
    url: https://github.com/cyware/ssi-service/issues
  description: 'The Self Sovereign Identity Service: Managing DIDs, Verifiable Credentials,
    and more!'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: 0.0.3
paths:
...
  /v1/schemas:
    get:
      consumes:
      - application/json
      description: List Credential Schemas stored by the service
      parameters:
      - description: Hint to the server of the maximum elements to return. More may
          be returned. When not set, the server will return all elements.
        in: query
        name: pageSize
        type: number
      - description: Used to indicate to the server to return a specific page of the
          list results. Must match a previous requests' `nextPageToken`.
        in: query
        name: pageToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.ListSchemasResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List Credential Schemas
      tags:
      - Schemas
    put:
      consumes:
      - application/json
      description: Create a schema for use with a Verifiable Credential
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateSchemaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_server_router.CreateSchemaResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a Credential Schema
      tags:
      - Schemas
  /v1/schemas/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Credential Schema by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a Credential Schema
      tags:
      - Schemas
    get:
      consumes:
      - application/json
      description: Get a Credential Schema by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.GetSchemaResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get a Credential Schema
      tags:
      - Schemas
  /v1/webhooks:
    get:
      consumes:
      - application/json
      description: Lists all webhooks stored by the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.ListWebhooksResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List webhooks
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: Creates a webhook
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_server_router.CreateWebhookResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a webhook
      tags:
      - Webhooks
  /v1/webhooks/{noun}/{verb}:
    get:
      consumes:
      - application/json
      description: Get a webhook by its ID
      parameters:
      - description: noun
        in: path
        name: noun
        required: true
        type: string
      - description: verb
        in: path
        name: verb
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.ListWebhookResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get a webhook
      tags:
      - Webhooks
  /v1/webhooks/{noun}/{verb}/{url}:
    delete:
      consumes:
      - application/json
      description: Delete a webhook by its ID
      parameters:
      - description: noun
        in: path
        name: noun
        required: true
        type: string
      - description: verb
        in: path
        name: verb
        required: true
        type: string
      - description: url
        in: path
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a webhook
      tags:
      - Webhooks
...
swagger: "2.0"